updated main file and added debug mode

+++ stab/src/stab_sched.h
@@ -1,3 +1,4 @@
+#include <stdint.h>
 /*
   Header file used as an interface for the scheduler
 
@@ -7,7 +8,6 @@
 
 
 
-int stab_init(void);
+int16_t stab_init(void);
 
-
-int stab_run(void);
+int16_t stab_run(void);
